+++ src/app/app.component.ts
@@ -3,8 +3,7 @@ import { FormControl } from '@angular/forms';
 type SN = string | number;
 @Component({
     selector: 'app-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.css']
+    templateUrl: './app.component.html'
 })
 export class AppComponent {
     public phone: number = 123456789;
@@ -20,6 +19,7 @@ export class AppComponent {
     public repeatForm: FormControl;
     public separatorForm: FormControl;
     public percent: FormControl;
+    public formSecureInput: FormControl;
 
     public pattern: pattern = {
         P: {
@@ -30,6 +30,7 @@ export class AppComponent {
     public numberOrStringFormModel: SN = '';
     public clearIfNotMatchModel: SN = '';
     public formModelDate: SN = '';
+    public secureMask: SN = '';
     public cpfModel: SN = '';
     public dateModel: SN = '';
     public showMaskModel: SN = '';
@@ -54,6 +55,7 @@ export class AppComponent {
         this.form = new FormControl('');
         this.form1 = new FormControl('');
         this.formDate = new FormControl('');
+        this.formSecureInput = new FormControl('');
         this.cpfFormControl = new FormControl();
         this.clearIfNotMatch = new FormControl();
         this.numberOrStringForm = new FormControl();