+++ src/app/ngx-mask/mask-applier.service.ts
@@ -4,28 +4,31 @@ import { config, IConfig } from './config';
 @Injectable()
 export class MaskApplierService {
     public dropSpecialCharacters: IConfig['dropSpecialCharacters'];
+    public hiddenInput: IConfig['hiddenInput'];
     public showTemplate!: IConfig['showTemplate'];
     public clearIfNotMatch!: IConfig['clearIfNotMatch'];
     public maskExpression: string = '';
+    public actualValue: string = '';
     public shownMaskExpression: string = '';
     public maskSpecialCharacters!: IConfig['specialCharacters'];
     public maskAvailablePatterns!: IConfig['patterns'];
     public prefix!: IConfig['prefix'];
     public sufix!: IConfig['sufix'];
     public customPattern!: IConfig['patterns'];
+    protected prevResult: string = '';
+    protected prevActualResult: string = '';
 
     private _shift!: Set<number>;
 
     public constructor(@Inject(config) protected _config: IConfig) {
         this._shift = new Set();
-        this.maskSpecialCharacters = this._config!.specialCharacters;
-        this.maskAvailablePatterns = this._config.patterns;
         this.clearIfNotMatch = this._config.clearIfNotMatch;
         this.dropSpecialCharacters = this._config.dropSpecialCharacters;
         this.maskSpecialCharacters = this._config!.specialCharacters;
         this.maskAvailablePatterns = this._config.patterns;
         this.prefix = this._config.prefix;
         this.sufix = this._config.sufix;
+        this.hiddenInput = this._config.hiddenInput;
     }
     // tslint:disable-next-line:no-any
     public applyMaskWithPattern(inputValue: string, maskAndPattern: [string, IConfig['patterns']]): string {
@@ -55,7 +58,7 @@ export class MaskApplierService {
             if (inputValue.match('[a-z]|[A-Z]') || inputValue.match(/[-!$%^&*()_+|~=`{}\[\]:";'<>?,\/]/)) {
                 inputValue = this._checkInput(inputValue);
                 if (inputValue.length >= 3 && inputValue !== '100') {
-                    inputValue = inputValue.substring(0, 2);
+                    inputValue = inputValue.substring(0, 5);
                 }
             }
             if (this.percentage(inputValue)) {
@@ -70,12 +73,19 @@ export class MaskApplierService {
             maskExpression === 'comma_separator' ||
             maskExpression.startsWith('comma_separator')
         ) {
-            if (inputValue.match('[a-z]|[A-Z]') || inputValue.match(/[-@#!$%^&*()_+|~=`{}\[\]:";<>?\/]/)) {
+            if (inputValue.match('[a-z]|[A-Z]') || inputValue.match(/[-@#!$%^&*()_+|~=`{}\[\]:";<>.?\/]/)) {
                 inputValue = this._checkInput(inputValue);
             }
             const precision: number = this.getPrecision(maskExpression);
             let strForSep: string;
             if (maskExpression.startsWith('dot_separator')) {
+                if (
+                    inputValue.indexOf('.') !== -1 &&
+                    inputValue.indexOf('.') === inputValue.lastIndexOf('.') &&
+                    inputValue.indexOf('.') > 3
+                ) {
+                    inputValue = inputValue.replace('.', ',');
+                }
                 inputValue =
                     inputValue.length > 1 && inputValue[0] === '0' && inputValue[1] !== ','
                         ? inputValue.slice(1, inputValue.length)
@@ -167,7 +177,12 @@ export class MaskApplierService {
                 ) {
                     result += inputSymbol;
                     cursor += 3;
-                } else if (this._checkSymbolMask(inputSymbol, maskExpression[cursor])) {
+                } else if (
+                    this._checkSymbolMask(inputSymbol, maskExpression[cursor]) ||
+                    (this.hiddenInput &&
+                        this.maskAvailablePatterns[maskExpression[cursor]] &&
+                        this.maskAvailablePatterns[maskExpression[cursor]].symbol === inputSymbol)
+                ) {
                     if (maskExpression[cursor] === 'H') {
                         if (Number(inputSymbol) > 2) {
                             result += 0;
@@ -268,6 +283,14 @@ export class MaskApplierService {
                 ) {
                     cursor += 3;
                     result += inputSymbol;
+                } else if (
+                    this.maskExpression[cursor + 1] === '?' &&
+                    this._findSpecialChar(this.maskExpression[cursor + 2]) &&
+                    this._findSpecialChar(inputSymbol) === this.maskExpression[cursor + 2] &&
+                    multi
+                ) {
+                    cursor += 3;
+                    result += inputSymbol;
                 }
             }
         }
@@ -301,7 +324,7 @@ export class MaskApplierService {
         return symbol;
     }
 
-    private _checkSymbolMask(inputSymbol: string, maskSymbol: string): boolean {
+    protected _checkSymbolMask(inputSymbol: string, maskSymbol: string): boolean {
         this.maskAvailablePatterns = this.customPattern ? this.customPattern : this.maskAvailablePatterns;
         return (
             this.maskAvailablePatterns[maskSymbol] &&
@@ -362,7 +385,7 @@ export class MaskApplierService {
     private _checkInput(str: string): string {
         return str
             .split('')
-            .filter((i: string) => i.match('\\d'))
+            .filter((i: string) => i.match('\\d') || i === '.' || i === ',')
             .join('');
     }
 }