+++ src/app/ngx-mask/test/add-suffix.spec.ts
@@ -1,0 +1,26 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NgxMaskModule } from '../ngx-mask.module';
+import { ReactiveFormsModule } from '@angular/forms';
+import { TestMaskComponent } from './utils/test-component.component';
+import { equal } from './utils/test-functions.component';
+
+describe('Directive: Mask', () => {
+    let fixture: ComponentFixture<TestMaskComponent>;
+    let component: TestMaskComponent;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [TestMaskComponent],
+            imports: [ReactiveFormsModule, NgxMaskModule.forRoot()]
+        });
+        fixture = TestBed.createComponent(TestMaskComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should have a suffix', () => {
+        component.mask = '00-000-000-00';
+        component.sufix = '$';
+        equal('6', '6$', fixture);
+    });
+});