+++ src/app/ngx-mask/test/utils/test-component.component.ts
@@ -16,6 +16,7 @@ import { IConfig } from 'public_api';
             [prefix]="prefix"
             [formControl]="form"
             [showMaskTyped]="showMaskTyped"
+            [hiddenInput]="hiddenInput"
             [(ngModel)]="ngModelValue"
         />
     `
@@ -31,4 +32,5 @@ export class TestMaskComponent {
     public sufix: IConfig['sufix'] = '';
     public specialCharacters!: IConfig['specialCharacters'];
     public showMaskTyped: IConfig['showMaskTyped'] = false;
+    public hiddenInput: IConfig['hiddenInput'] = false;
 }